feat: return event details with active coupons

+++ src/main/java/com/eyetech/events/dtos/CuponRequestDTO.java
@@ -1,7 +1,8 @@
 package com.eyetech.events.dtos;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
-public record CuponRequestDTO(String code, Integer discount, Date valid) {
+public record CuponRequestDTO(String code, BigDecimal discount, Date valid) {
     
 }
